If value is null, set to empty
Original method was to bypass replacement where token variable is not yet at replacement time. This has been reconsidered, and instead we'll replace with an empty value for these cases. This will prevent token strings from being passed in the final product of replacement; as the most common case is URL replacement, this will result in empty parameters if the variable hasn't been set.